backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
supabase==2.3.0
asyncpg==0.29.0
redis==5.0.1
httpx==0.26.0
python-dotenv==1.0.0
tenacity==8.2.3
modal==0.56.4374
openai==1.6.1
unstructured==0.11.2
numpy==1.26.3
sentence-transformers==2.2.2
langchain==0.1.0
langgraph==0.0.17
sentry-sdk==1.39.1
structlog==24.1.0 